AuthService/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-dotenv
gunicorn
psycopg2-binary
django-dotenv
pytest-django
pytest
requests
django-cors-headers